dimensionOS/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Login==0.6.3
Flask-Session==0.5.0
authlib==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
cryptography==41.0.7

